requirements.txt
numpy~=1.26.2
Pillow~=8.4.0
matplotlib~=3.5.1
pynput~=1.7.6
opencv-python==4.8.0.74
opencv-contrib-python
omegaconf~=2.3.0
PyYAML~=6.0
torchvision~=0.14.0+cu117
scipy~=1.11.4
hydra-core~=1.3.2
torch~=1.13.0+cu117
